+++ pixeltable/catalog/table_version.py
@@ -25,7 +25,7 @@ from pixeltable.runtime import get_runtime
 from pixeltable.utils.object_stores import ObjectOps
 
 from .column import Column
-from .globals import _ROWID_COLUMN_NAME, IndexSpec, MediaValidation, QColumnId, TableVersionKey, is_valid_identifier
+from .globals import _ROWID_COLUMN_NAME, IndexSpec, MediaValidation, is_valid_identifier
 from .tbl_ops import (
     CreateColumnMdOp,
     CreateStoreColumnsOp,
@@ -37,7 +37,7 @@ from .tbl_ops import (
     TableOp,
     TableOpsBuilder,
 )
-from .types import TableVersionMd
+from .types import QColumnId, TableVersionKey, TableVersionMd
 from .update_status import RowCountStats, UpdateStatus
 from .utils import generate_idx_name, validate_idxs
