+++ pixeltable/catalog/table_version_handle.py
@@ -7,8 +7,8 @@ from uuid import UUID
 from pixeltable import exceptions as excs
 from pixeltable.runtime import get_runtime
 
-from .globals import TableVersionKey
 from .table_version import TableVersion
+from .types import TableVersionKey
 
 if TYPE_CHECKING:
     from pixeltable.catalog import Column
